Time: 52 ms (7.47%), Space: 103.7 MB (29.49%) - LeetHub

+++ 0234-palindrome-linked-list/0234-palindrome-linked-list.java
@@ -18,7 +18,7 @@ class Solution {
             temp = temp.next;
         }
         while (head != null) {
-            if (head.val != (int) stack.pop()) {
+            if (head.val != stack.pop()) {
                 return false;
             }
             head = head.next;
